JJP-142: Create RSVP create form and page
