+++ app/frontend/src/components/RSVPCreateComponent/RSVPCreatePageStyles.ts
@@ -1,0 +1,37 @@
+export const RSVPCreateComponentStyles = {
+  fieldBoxStyles: {
+    alignItems: "center",
+    margin: "auto",
+    boxShadow: 3,
+    bgcolor:"var(--rsvp-create-page-bgcolor)"
+  },
+
+  formBoxStyles: {
+    display: "flex",
+    gap: 4,
+  },
+
+  firstInputGroupStyles: {
+    display: "flex",
+    flexDirection: "column",
+    margin: "auto",
+    gap: 2,
+  },
+
+  secondInputGroupStyles: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    gap: 3,
+    width: "60%",
+    margin: "auto",
+    justifyContent: "center",
+    bgcolor: "var(--rsvp-create-page-form)",
+  },
+
+  buttonStyles: {
+    width: "5%",
+    margin: "auto",
+  },
+  
+};
